Fix series logos asset paths
